table added

+++ src/api/routes.py
@@ -2,7 +2,7 @@
 This module takes care of starting the API Server, Loading the DB and Adding the endpoints
 """
 from flask import Flask, request, jsonify, url_for, Blueprint
-from api.models import db, User, Menu, Restaurant
+from api.models import db, User, Menu, Restaurant, Table
 from api.utils import generate_sitemap, APIException
 from flask_cors import CORS
 
@@ -37,22 +37,117 @@ def add_restaurant():
     db.session.commit()
     return jsonify(new_restaurant.serialize()), 201
 
+@api.route('/restaurants/<int:restaurant_id>/tables', methods=['POST'])
+def add_table(restaurant_id):
+    data = request.get_json()
+    new_table = Table(restaurant_id=restaurant_id, number=data['number'])
+    db.session.add(new_table)
+    db.session.commit()
+    return jsonify(new_table.serialize()), 201
 
-@api.route('/restaurants/<int:restaurant_id>/menu', methods=['GET'])
-def get_menu(restaurant_id):
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>', methods=['GET'])
+def get_table(restaurant_id, table_id):
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if table:
+        return jsonify(table.serialize()), 200
+    return jsonify({"error": "Table not found"}), 404
+
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>', methods=['PUT'])
+def update_table(restaurant_id, table_id):
+    data = request.get_json()
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if table:
+        table.number = data['number']
+        db.session.commit()
+        return jsonify(table.serialize()), 200
+    return jsonify({"error": "Table not found"}), 404
+
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>', methods=['DELETE'])
+def delete_table(restaurant_id, table_id):
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if table:
+        db.session.delete(table)
+        db.session.commit()
+        return jsonify({"message": "Table deleted"}), 200
+    return jsonify({"error": "Table not found"}), 404
+
+# @api.route('/restaurants/<int:restaurant_id>/menu', methods=['GET'])
+# def get_menu(restaurant_id):
+#     restaurant = Restaurant.query.get(restaurant_id)
+#     if not restaurant:
+#         return jsonify({"error": "Restaurant not found"}), 404
+
+#     menu_items = Menu.query.filter_by(restaurant_id=restaurant_id).all()
+#     return jsonify([item.serialize() for item in menu_items]), 200
+
+
+# @api.route('/restaurants/<int:restaurant_id>/menu', methods=['POST'])
+# def add_menu_item(restaurant_id):
+#     restaurant = Restaurant.query.get(restaurant_id)
+#     if not restaurant:
+#         return jsonify({"error": "Restaurant not found"}), 404
+
+#     data = request.json
+#     new_menu_item = Menu(
+#         name=data['name'],
+#         description=data['description'],
+#         price=data['price'],
+#         category=data['category'],
+#         restaurant_id=restaurant_id
+#     )
+#     db.session.add(new_menu_item)
+#     db.session.commit()
+#     return jsonify(new_menu_item.serialize()), 201
+
+
+# @api.route('/restaurants/<int:restaurant_id>/menu/<int:item_id>', methods=['PUT'])
+# def update_menu_item(restaurant_id, item_id):
+#     menu_item = Menu.query.filter_by(id=item_id, restaurant_id=restaurant_id).first()
+#     if not menu_item:
+#         return jsonify({"error": "Menu item not found"}), 404
+
+#     data = request.json
+#     menu_item.name = data.get('name', menu_item.name)
+#     menu_item.description = data.get('description', menu_item.description)
+#     menu_item.price = data.get('price', menu_item.price)
+#     menu_item.category = data.get('category', menu_item.category)
+
+#     db.session.commit()
+#     return jsonify(menu_item.serialize()), 200
+
+
+# @api.route('/restaurants/<int:restaurant_id>/menu/<int:item_id>', methods=['DELETE'])
+# def delete_menu_item(restaurant_id, item_id):
+#     menu_item = Menu.query.filter_by(id=item_id, restaurant_id=restaurant_id).first()
+#     if not menu_item:
+#         return jsonify({"error": "Menu item not found"}), 404
+
+#     db.session.delete(menu_item)
+#     db.session.commit()
+#     return jsonify({"message": "Menu item deleted successfully"}), 200
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>/menu', methods=['GET'])
+def get_menu(restaurant_id, table_id):
     restaurant = Restaurant.query.get(restaurant_id)
     if not restaurant:
         return jsonify({"error": "Restaurant not found"}), 404
+
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if not table:
+        return jsonify({"error": "Table not found"}), 404
 
     menu_items = Menu.query.filter_by(restaurant_id=restaurant_id).all()
     return jsonify([item.serialize() for item in menu_items]), 200
 
 
-@api.route('/restaurants/<int:restaurant_id>/menu', methods=['POST'])
-def add_menu_item(restaurant_id):
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>/menu', methods=['POST'])
+def add_menu_item(restaurant_id, table_id):
     restaurant = Restaurant.query.get(restaurant_id)
     if not restaurant:
         return jsonify({"error": "Restaurant not found"}), 404
+
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if not table:
+        return jsonify({"error": "Table not found"}), 404
 
     data = request.json
     new_menu_item = Menu(
@@ -67,11 +162,15 @@ def add_menu_item(restaurant_id):
     return jsonify(new_menu_item.serialize()), 201
 
 
-@api.route('/restaurants/<int:restaurant_id>/menu/<int:item_id>', methods=['PUT'])
-def update_menu_item(restaurant_id, item_id):
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>/menu/<int:item_id>', methods=['PUT'])
+def update_menu_item(restaurant_id, table_id, item_id):
     menu_item = Menu.query.filter_by(id=item_id, restaurant_id=restaurant_id).first()
     if not menu_item:
         return jsonify({"error": "Menu item not found"}), 404
+
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if not table:
+        return jsonify({"error": "Table not found"}), 404
 
     data = request.json
     menu_item.name = data.get('name', menu_item.name)
@@ -83,11 +182,15 @@ def update_menu_item(restaurant_id, item_id):
     return jsonify(menu_item.serialize()), 200
 
 
-@api.route('/restaurants/<int:restaurant_id>/menu/<int:item_id>', methods=['DELETE'])
-def delete_menu_item(restaurant_id, item_id):
+@api.route('/restaurants/<int:restaurant_id>/tables/<int:table_id>/menu/<int:item_id>', methods=['DELETE'])
+def delete_menu_item(restaurant_id, table_id, item_id):
     menu_item = Menu.query.filter_by(id=item_id, restaurant_id=restaurant_id).first()
     if not menu_item:
         return jsonify({"error": "Menu item not found"}), 404
+
+    table = Table.query.filter_by(id=table_id, restaurant_id=restaurant_id).first()
+    if not table:
+        return jsonify({"error": "Table not found"}), 404
 
     db.session.delete(menu_item)
     db.session.commit()
